refactor: 사용자 책장 템플릿 으로 분리
